pam crate: add tests.
I've put the tests in bin/main.rs instead of in lib.rs,
because "cargo test" for the library links the tests without -lpam,
so it fails. It sucks, because now the test for the magic service
"xyzzy-test-test" is always present in src/pam.rs, instead of only
during tests.